[FIX] iwp_website: Currency symbol on manual share
