+++ iwp_website/controllers/investor_portal.py
@@ -12,23 +12,9 @@ from odoo.http import request
 from odoo.tools.translate import _
 
 from .user_form import InvestorCompanyForm, InvestorPersonForm
+from .tools import monetary_to_text
 
 # TODO: Try to not give sudo object to a view.
-
-
-def monetary_to_text(value, currency=None):
-    if currency is None:
-        currency = (
-            request.env['res.company']._company_default_get().currency_id
-        )
-    value_to_html = request.env['ir.qweb.field.monetary'].value_to_html
-    html_val = value_to_html(float(value), {"display_currency": currency})
-    raw_val = (
-        html_val
-        .replace('<span class="oe_currency_value">', "")
-        .replace("</span>", "")
-    )
-    return raw_val
 
 
 class InvestorPortal(CustomerPortal):
